# docs/openapi.yaml
openapi: 3.2.0
info:
  title: kpblog API
  version: 0.1.0
  description: |
    kpblog 后端 JSON API 契约。

    所有 `/api/*` JSON 接口统一使用 envelope:

    - 成功: `{ "success": true, "data": ... }`
    - 失败: `{ "success": false, "error": { "code": "...", "message": "..." } }`

    Web 浏览器通过同域 `/api/*` 访问；Next.js 将请求代理到独立 Fastify
    进程。`/rss.xml` 等由 Web 输出的非 JSON 内容不包含在本文件中。
servers:
  - url: http://localhost:3001
    description: Local same-origin Web proxy
  - url: http://127.0.0.1:3002
    description: Local Fastify API (private)
  - url: https://kpblog.cc
    description: Production
...
paths:
...
  /api/public/articles/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Public]
      summary: Get a published article by slug
      responses:
        "200":
          description: Published article or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
  /api/articles:
    get:
      tags: [Articles]
      summary: List articles
      description: |
        Public requests return published articles only.
        Authenticated admin requests can pass `published=all` or `published=draft`.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag slug.
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category slug.
        - name: published
          in: query
          schema:
            type: string
            enum: [all, draft, published]
          description: Admin-only publication filter.
        - name: q
          in: query
          schema:
            type: string
            maxLength: 100
          description: >-
            Case-insensitive keyword filter over title, excerpt, content and tag
            names. The API truncates the value to 100 characters.
      responses:
        "200":
          description: Paginated articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
    post:
      tags: [Articles]
      summary: Create article
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleCreateInput"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "201":
          description: Article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/articles/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Articles]
      summary: Get article by id
      description: Draft articles require an authenticated admin session.
      responses:
        "200":
          description: Article detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Articles]
      summary: Update article
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdateInput"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Article updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Articles]
      summary: Delete article
      security:
        - sessionCookie: []
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Article deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/comments:
    get:
      tags: [Comments]
      summary: List comments
      description: |
        With `postId`, returns approved public comment threads for that post.
        Without `postId`, requires a current database-backed admin session and returns the moderation list.
      parameters:
        - name: postId
          in: query
          schema:
            type: string
        - name: approved
          in: query
          schema:
            type: string
            enum: [pending, approved, "true", "false"]
          description: Admin moderation filter. Ignored when `postId` is provided.
        - name: scope
          in: query
          schema:
            type: string
            enum: [post, guestbook]
          description: >-
            Admin-only source filter. `post` returns article comments, `guestbook`
            returns guestbook entries. Any other value means no filtering.
            Ignored when `postId` is provided.
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Public comment threads or admin comment list
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CommentListResponse"
                  - $ref: "#/components/schemas/AdminCommentListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Comments]
      summary: Submit comment
      description: New comments are created with `pendingReview=true` and are hidden until approved.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreateInput"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "201":
          description: Comment submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/guestbook:
    post:
      tags: [Comments]
      summary: Submit a guestbook entry
      description: >-
        Same validation and moderation queue as article comments, but the entry
        is not attached to any post. Any `postId` in the body is ignored —
        ownership is decided by the endpoint. Rate limit is counted separately
        from article comments (10 per 10 minutes per IP).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestbookEntryCreateInput"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "201":
          description: Entry submitted for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/comments/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    put:
      tags: [Comments]
      summary: Approve or reject comment
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentModerationInput"
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Comment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCommentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Comments]
      summary: Delete comment
      security:
        - sessionCookie: []
      responses:
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: kp API key
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    TooManyRequests:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
...
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
  schemas:
    ApiFailure:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ApiError"
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [BAD_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, TOO_MANY_REQUESTS, INTERNAL_ERROR]
        message:
          type: string
        retryAfterSeconds:
          type: integer
          minimum: 1
          description: >-
            Only present on rate-limit (429) responses: seconds until the caller may
            try again. An ADMIN account allows exactly 3 password attempts: attempts
            1-3 return 401, the 4th is refused before password verification and locks
            the account for 15 minutes (configurable via ADMIN_LOGIN_LOCKOUT_MINUTES).
            Regular accounts allow 10 failures per 15 minutes. The lock follows the
            account, not the IP, so rotating IPs does not clear it.
...
    Envelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data: {}
    DeleteResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [deleted]
              properties:
                deleted:
                  type: boolean
                  const: true
...
    TagRef:
      type: object
      required: [name, slug]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
    CategoryRef:
      type:
        - object
        - "null"
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
    PostSummary:
      type: object
      required: [id, slug, title, excerpt, coverImage, published, publishedAt, createdAt, updatedAt, category, tags]
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        excerpt:
          type:
            - string
            - "null"
        coverImage:
          type:
            - string
            - "null"
        published:
          type: boolean
        publishedAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          $ref: "#/components/schemas/CategoryRef"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagRef"
    PostDetail:
      allOf:
        - $ref: "#/components/schemas/PostSummary"
        - type: object
          required: [content, categoryId]
          properties:
            content:
              type: string
            categoryId:
              type:
                - string
                - "null"
...
    ArticleListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [items, total, page, pageSize, totalPages]
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/PostSummary"
                total:
                  type: integer
                page:
                  type: integer
                pageSize:
                  type: integer
                totalPages:
                  type: integer
    PostDetailResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PostDetail"
    ArticleCreateInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type:
            - string
            - "null"
        content:
          type: string
        coverImage:
          type:
            - string
            - "null"
        published:
          type: boolean
          default: false
        publishedAt:
          type:
            - string
            - "null"
          description: >-
            ISO date-time. Lets you backdate a post so it files under the right
            year in /archive and sorts correctly in listings. Omit to keep the
            current behaviour (now, when the post is published); send null to
            clear it. Ignored for drafts, which have no publish date.
        categoryId:
          type:
            - string
            - "null"
        projectId:
          type:
            - string
            - "null"
          description: >-
            Optional project collection (Post.projectId). Omit to keep the
            existing value; send null to remove the post from its project.
        tagIds:
          type: array
          items:
            type: string
    ArticleUpdateInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type:
            - string
            - "null"
        content:
          type: string
        coverImage:
          type:
            - string
            - "null"
        published:
          type: boolean
        publishedAt:
          type:
            - string
            - "null"
          description: >-
            ISO date-time. Lets you backdate a post so it files under the right
            year in /archive and sorts correctly in listings. Omit to keep the
            existing date; send null to clear it. Ignored for drafts, which have
            no publish date.
        categoryId:
          type:
            - string
            - "null"
        projectId:
          type:
            - string
            - "null"
          description: >-
            Optional project collection (Post.projectId). Omit to keep the
            existing value; send null to remove the post from its project.
        tagIds:
          type: array
          items:
            type: string
    CommentReply:
      type: object
      required: [id, author, content, createdAt]
      properties:
        id:
          type: string
        author:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CommentThread:
      allOf:
        - $ref: "#/components/schemas/CommentReply"
        - type: object
          required: [replies]
          properties:
            replies:
              type: array
              items:
                $ref: "#/components/schemas/CommentReply"
    CommentListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [items, total, page, pageSize, totalPages]
              description: >-
                Top-level comment threads are paginated. Replies are returned in
                full inside each thread and are not paginated separately.
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/CommentThread"
                total:
                  type: integer
                page:
                  type: integer
                pageSize:
                  type: integer
                totalPages:
                  type: integer
    CommentCreateInput:
      type: object
      required: [postId, content]
      properties:
        postId:
          type: string
        parentId:
          type:
            - string
            - "null"
        author:
          type:
            - string
            - "null"
        email:
          type:
            - string
            - "null"
          format: email
        content:
          type: string
          maxLength: 2000
    GuestbookEntryCreateInput:
      type: object
      required: [content]
      properties:
        parentId:
          type:
            - string
            - "null"
          description: Must reference an approved top-level guestbook entry.
        author:
          type:
            - string
            - "null"
          description: Ignored for signed-in users, who comment under their account identity.
        email:
          type:
            - string
            - "null"
          format: email
        content:
          type: string
          maxLength: 2000
    CommentCreateResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [id, pendingReview]
              properties:
                id:
                  type: string
                pendingReview:
                  type: boolean
                  const: true
    CommentModerationInput:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
...
    AdminComment:
      type: object
      required: [id, author, email, content, approved, createdAt, postId, parentId, scope, post, parent]
      properties:
        id:
          type: string
        author:
          type: string
        email:
          type:
            - string
            - "null"
        content:
          type: string
        approved:
          type: boolean
        createdAt:
          type: string
          format: date-time
        postId:
          type:
            - string
            - "null"
          description: Null for guestbook entries.
        parentId:
          type:
            - string
            - "null"
        scope:
          type: string
          enum: [post, guestbook]
          description: Whether this came from an article or the guestbook.
        post:
          type:
            - object
            - "null"
          required: [id, title, slug]
          properties:
            id:
              type: string
            title:
              type: string
            slug:
              type: string
        parent:
          type:
            - object
            - "null"
          required: [id, author]
          properties:
            id:
              type: string
            author:
              type: string
    AdminCommentResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AdminComment"
    AdminCommentListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [items, total, page, pageSize, totalPages]
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/AdminComment"
                total:
                  type: integer
                page:
                  type: integer
                pageSize:
                  type: integer
                totalPages:
                  type: integer
...